Fixed broken doxygen generation for Plugin Framework
In some of the header files in the plugin framework, the two macros

  Q_DECLARE_PRIVATE(ctkPlugin)
  Q_DISABLE_COPY(ctkPlugin)

appeared at the beginning of the class declaration. This is valid c++,
but the doxygen generator intrepret this as a private statement and
thus all subsequent public method declarations are interpreted as private,
thus hiding them from the user documentation.

We have no idea why doxygen does this.

Investigation was done in cooperation with @jcfr.

Fix: Moved the offending macros to the end of the class declaration.