BUG: Prevent infinite recursion in Qt log handler
Slicer sometimes crashed on exit because in Qt log handler in
ctk::qtHandleToString (in ctkUtils) while executing line
"QTextStream s(&str)" a QDeviceClosedNotifier::QDeviceClosedNotifier()
object was  instantiated that tried to log a message, which eventually
got to ctk::qtHandleString, creating an infinite recursion.

To solve this, a counter is added that is incremented at each
recursion level and when it reaches a threshold then it does not try
to handle the message, thus preventing the crash.

Closes #543

Reviewed-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>