Fixed again duplicate signal connection
Cause of the bug:
If the dicom database is passed via the ctor then the ctor called
setDicomDatabase, which called setUpTableview, where the signals were
connected. Afterwards init was called by the ctor, which also called
setDicomDatabase, which then lead to the duplicate connection.

Solution:
The setUpTableview method is not really needed. All the connections can be
done in the init function. Hence I removed the setUpTableview function.